chore: minor code refactor
