+++ apps/tangle-dapp/app/nomination/[validatorAddress]/InfoCard.tsx
@@ -9,36 +9,27 @@ import {
   Typography,
 } from '@webb-tools/webb-ui-components';
 import { shortenString } from '@webb-tools/webb-ui-components/utils/shortenString';
-import { FC, useEffect, useMemo, useState } from 'react';
+import { FC } from 'react';
 import { twMerge } from 'tailwind-merge';
 
 import { SocialChip, TangleCard } from '../../../components';
 import useNetworkStore from '../../../context/useNetworkStore';
-import useRestakingRoleLedger from '../../../data/restaking/useRestakingRoleLedger';
-import useCurrentEra from '../../../data/staking/useCurrentEra';
+import useValidatorInfoCard from '../../../data/validatorDetails/useValidatorInfoCard';
+import useFormatNativeTokenAmount from '../../../hooks/useFormatNativeTokenAmount';
 import { ExplorerType } from '../../../types';
-import {
-  extractDataFromIdentityInfo,
-  formatTokenBalance,
-  IdentityDataType,
-} from '../../../utils/polkadot';
-import { getPolkadotApiPromise } from '../../../utils/polkadot/api';
-import {
-  getProfileTypeFromRestakeRoleLedger,
-  getTotalRestakedFromRestakeRoleLedger,
-} from '../../../utils/polkadot/restake';
 import ValueSkeleton from './ValueSkeleton';
 
-interface ValidatorBasicInfoCardProps {
+interface InfoCardProps {
   validatorAddress: string;
   className?: string;
 }
 
-const ValidatorBasicInfoCard: FC<ValidatorBasicInfoCardProps> = ({
+const InfoCard: FC<InfoCardProps> = ({
   validatorAddress,
   className,
-}: ValidatorBasicInfoCardProps) => {
-  const { network, nativeTokenSymbol, rpcEndpoint } = useNetworkStore();
+}: InfoCardProps) => {
+  const { network, rpcEndpoint } = useNetworkStore();
+  const formatNativeTokenAmount = useFormatNativeTokenAmount();
 
   const {
     name,
@@ -50,7 +41,7 @@ const ValidatorBasicInfoCard: FC<ValidatorBasicInfoCardProps> = ({
     email,
     web,
     isLoading,
-  } = useValidatorBasicInfo(rpcEndpoint, validatorAddress);
+  } = useValidatorInfoCard(rpcEndpoint, validatorAddress);
 
   return (
     <TangleCard className={twMerge('min-h-[300px]', className)}>
@@ -123,7 +114,7 @@ const ValidatorBasicInfoCard: FC<ValidatorBasicInfoCardProps> = ({
                   className="whitespace-nowrap"
                 >
                   {totalRestaked
-                    ? formatTokenBalance(totalRestaked, nativeTokenSymbol)
+                    ? formatNativeTokenAmount(totalRestaked)
                     : '--'}
                 </Typography>
               )}
@@ -162,99 +153,4 @@ const ValidatorBasicInfoCard: FC<ValidatorBasicInfoCardProps> = ({
   );
 };
 
-export default ValidatorBasicInfoCard;
-
-function useValidatorBasicInfo(rpcEndpoint: string, validatorAddress: string) {
-  const { data: currentEra } = useCurrentEra();
-  const { data: ledgerOpt, isLoading: isLoadingLedgerOpt } =
-    useRestakingRoleLedger(validatorAddress);
-
-  const [name, setName] = useState<string | null>(null);
-  const [email, setEmail] = useState<string | null>(null);
-  const [web, setWeb] = useState<string | null>(null);
-  const [twitter, setTwitter] = useState<string | null>(null);
-  const [nominations, setNominations] = useState<number | null>(null);
-  const [isActive, setIsActive] = useState<boolean | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  const restakingMethod = useMemo(
-    () => getProfileTypeFromRestakeRoleLedger(ledgerOpt),
-    [ledgerOpt]
-  );
-
-  const totalRestaked = useMemo(
-    () => getTotalRestakedFromRestakeRoleLedger(ledgerOpt),
-    [ledgerOpt]
-  );
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const api = await getPolkadotApiPromise(rpcEndpoint);
-      const fetchNameAndSocials = async () => {
-        const identityData = await api.query.identity.identityOf(
-          validatorAddress
-        );
-
-        if (identityData.isSome) {
-          const identity = identityData.unwrap();
-          const info = identity[0]?.info;
-          if (info) {
-            setName(extractDataFromIdentityInfo(info, IdentityDataType.NAME));
-            setEmail(extractDataFromIdentityInfo(info, IdentityDataType.EMAIL));
-            setWeb(extractDataFromIdentityInfo(info, IdentityDataType.WEB));
-            const twitterName = extractDataFromIdentityInfo(
-              info,
-              IdentityDataType.TWITTER
-            );
-            setTwitter(
-              twitterName === null
-                ? null
-                : `https://twitter.com/${twitterName.substring(1)}`
-            );
-          }
-        }
-      };
-
-      const fetchNominations = async () => {
-        if (currentEra === null || !api.query.staking?.erasStakersOverview) {
-          setNominations(null);
-          setIsActive(null);
-          return;
-        }
-
-        const erasStakersOverviewData =
-          await api.query.staking.erasStakersOverview(
-            currentEra,
-            validatorAddress
-          );
-        if (erasStakersOverviewData.isSome) {
-          const nominatorCount =
-            erasStakersOverviewData.unwrap().nominatorCount;
-          setNominations(nominatorCount.toNumber());
-          setIsActive(true);
-          return;
-        }
-
-        setNominations(null);
-        setIsActive(false);
-      };
-
-      await Promise.all([fetchNameAndSocials(), fetchNominations()]);
-      setIsLoading(false);
-    };
-
-    fetchData();
-  }, [validatorAddress, rpcEndpoint, currentEra]);
-
-  return {
-    name,
-    isActive,
-    totalRestaked,
-    restakingMethod,
-    nominations,
-    twitter,
-    email,
-    web,
-    isLoading: isLoading || isLoadingLedgerOpt,
-  };
-}
+export default InfoCard;
