+++ apps/tangle-dapp/app/nomination/[validatorAddress]/page.tsx
@@ -1,10 +1,10 @@
 import { isAddress } from '@polkadot/util-crypto';
 import { notFound } from 'next/navigation';
 
+import InfoCard from './InfoCard';
 import NodeSpecificationsTable from './NodeSpecificationsTable';
 import RoleDistributionCard from './RoleDistributionCard';
 import ServiceTableTabs from './ServiceTableTabs';
-import ValidatorBasicInfoCard from './ValidatorBasicInfoCard';
 
 // TODO: might need to add metadata here
 
@@ -22,10 +22,7 @@ export default function ValidatorDetails({
   return (
     <div className="my-5 space-y-10">
       <div className="flex flex-col lg:flex-row gap-5 items-stretch">
-        <ValidatorBasicInfoCard
-          validatorAddress={validatorAddress}
-          className="flex-1"
-        />
+        <InfoCard validatorAddress={validatorAddress} className="flex-1" />
         <RoleDistributionCard
           validatorAddress={validatorAddress}
           className="flex-1"
